+++ internal/pkg/cmdimpl/list_test.go
@@ -14,7 +14,7 @@ func TestListImpl_1(t *testing.T) {
 	sourceDir := "../.."
 	searchRoot := "../../.."
 
-	if fileRepo, err := cmdimpl.ListImpl(sourceDir, searchRoot, blackList, whiteList, true, false); err != nil {
+	if fileRepo, err := cmdimpl.ListImpl(sourceDir, searchRoot, blackList, whiteList, true, false, "", false); err != nil {
 		t.Errorf("Error while searching for duplicates: %v", err)
 	} else {
 		fmt.Println("FileRepo.Size: ", fileRepo.Size())
@@ -39,7 +39,7 @@ func TestListImpl_2(t *testing.T) {
 	sourceDir := "../.."
 	searchRoot := "../../.."
 
-	if fileRepo, err := cmdimpl.ListImpl(sourceDir, searchRoot, blackList, whiteList, true, false); err != nil {
+	if fileRepo, err := cmdimpl.ListImpl(sourceDir, searchRoot, blackList, whiteList, true, false, "", false); err != nil {
 		t.Errorf("Error while searching for duplicates: %v", err)
 	} else {
 		if fileRepo.Size() != 1 {
@@ -70,7 +70,7 @@ func TestListImpl_3(t *testing.T) {
 	blackList := []string{".git"}
 	sourceDir := "../../.."
 
-	if fileRepo, err := cmdimpl.ListImpl(sourceDir, "", blackList, whiteList, true, true); err != nil {
+	if fileRepo, err := cmdimpl.ListImpl(sourceDir, "", blackList, whiteList, true, false, "", true); err != nil {
 		t.Errorf("Error while searching for duplicates: %v", err)
 	} else {
 		fmt.Println("FileRepo.Size: ", fileRepo.Size())
@@ -94,7 +94,7 @@ func TestListImpl_4(t *testing.T) {
 	sourceDir := "../.."
 	searchRoot := "../../.."
 
-	if fileRepo, err := cmdimpl.ListImpl(sourceDir, searchRoot, blackList, whiteList, true, true); err != nil {
+	if fileRepo, err := cmdimpl.ListImpl(sourceDir, searchRoot, blackList, whiteList, true, false, "", true); err != nil {
 		t.Errorf("Error while searching for duplicates: %v", err)
 	} else {
 		fmt.Println("FileRepo.Size: ", fileRepo.Size())