+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.3.0
+
+* adding 'outputFile' flag to list command
+* enable JSON output for list command
+
 # v0.2.0
 
 * handle overlapping files in situations where source dir and search root overlap